Enhance PlayerProfile and Game classes with additional fields and validation logic

+++ src/main/java/de/paulm/four_in_a_row/profil/PlayerStatistic.java
@@ -17,9 +17,9 @@ public class PlayerStatistic {
     @GeneratedValue
     private Long id;
 
-    private Long totalGames;
+    private int totalGames;
 
-    private Long gamesWon;
+    private int gamesWon;
 
     private LocalDate lastPlayedOn;
